+++ FluidNC/src/Motors/TMC5160Driver.h
@@ -10,11 +10,11 @@
 
 #include <cstdint>
 
-const float TMC2130_RSENSE_DEFAULT = 0.11f;
+const float TMC5160_RSENSE_DEFAULT = 0.075f;
 
 namespace MotorDrivers {
 
-    class TMC2130Driver : public TrinamicSpiDriver {
+    class TMC5160Driver : public TrinamicSpiDriver {
     public:
         // Overrides for inherited methods
         void init() override;
@@ -23,15 +23,21 @@ namespace MotorDrivers {
         void debug_message() override;
         void validate() const override { StandardStepper::validate(); }
 
+        void group(Configuration::HandlerBase& handler) override {
+            TrinamicSpiDriver::group(handler);
+            handler.item("tpfd", _tpfd, 0, 15);
+        }
+
         // Name of the configurable. Must match the name registered in the cpp file.
-        const char* name() const override { return "tmc_2130"; }
+        const char* name() const override { return "tmc_5160"; }
 
     private:
-        TMC2130Stepper* tmc2130 = nullptr;
+        TMC5160Stepper* tmc5160 = nullptr;
+
+        uint8_t _tpfd = 4;
 
         bool test();
-        void set_mode(bool isHoming);
-        void set_homing_mode();
+        void set_registers(bool isHoming);
         void trinamic_test_response();
         void trinamic_stepper_enable(bool enable);
     };
